Fixed a bug in txt2map that caused crash if targets/obstacles < 2
The function length() was used on targets and obstacles matrices
for returning the number of targets/obstacles, given by the number
of rows). The length() function gives the highest dimension of a
matrix, so it worked fine when there were at least two targets and
obstacles, but would fail with 1 osbtacle/target, as the matrices
in question always have 2 columns (x and y).
The function was replaced with size(M,1), which gives the
size of the 1st dimension of the matrix M, so the number of rows.