Updating SW model to work on 2016a, for easier comparisons to HW model
Changes to enable the SoftwareModel to run on Matlab 2016a, to
enable more direct comparisons to the hardware model.

The changes are mostly just direct replacements of functions which were
not compatible with 2016a to ones that are.
In the case of getScore.m, a bigger change was made, as the old
getScore function was changed to the updated version that was already
used on the HW model. Consequently, RunModel.m had to also be changed
to call this new version of the getScore function.