Changes to HW model: reporting works; different parameters to PIDs
The PID parameters for PIDs 2,3 and 4 (both X and Y) were the same.
Changed these to be different, so they can be set independently,
to use the values Kenny tested.

The reporting now gives the full set of time steps also in longer
simulation times. Previously it worked up to around 40 second
simulations, but would only give the last 1000 or 2000 datapoints
for longer simulations.