Updated the movement verification scripts; fixed a couple of bugs
Updated the movement verification scripts to work with the latest
model, and 2016a.
Also fixed a bug in the model that would cause an abort if the
path length is less than 2. This was due to incorrect use of
length(M) in a two-dimensional matrix M, which gives size(M,2),
while we actually wanted size(M,1). The error caused an abort only
when the path was very short, in this case less than 2.

The path_to_waypoints.m script had a bug causing an abort in some
cases of short paths. This is also fixed.