Enabled reversing the order of the path points, for verification
The path finding algorithm always starts a path with the target
closest to the origin. This was causing "return paths" -- paths
from a certain point back to (3,3) -- in verification being set
to the same as the paths from (3,3) to the given point. The change
allows setting a flag into RunModelVerification.m that reverses
the order of the path generated for the verification runs.